refactor: mentor 오타 수정
