+++ frontend/src/domain/booking/components/MentorInfoCard/MentorInfoCart.stories.tsx
@@ -2,7 +2,7 @@ import { ThemeProvider } from '@emotion/react';
 
 import { THEME } from '../../../../common/styles/theme';
 
-import MentoInfoCard from './MentoInfoCard';
+import MentoInfoCard from './MentorInfoCard';
 
 import type { Meta, StoryObj } from '@storybook/react-webpack5';
